Split pattern
I was astonished to see such an annoying wart around the use of a slice
as a split pattern. &['a', 'b', 'c'] really ought to work just fine, but
it doesn't because it is detected as a reference to an ARRAY rather than
as a slice. How annoying!

In any case, this works, and I found it less distracting than the
explicit slicing notation. Your mileage may vary.