Use byte string literal thingies
In Rust, a string prefaced with a b is actually a byte slice. Likewise,
a character prefaced with a b is instead a byte. For example, b"Hello,
World!" is a byte slice containing the ascii bytes for the string
"Hello, World!"; likewise, b'a' is the byte value of 'a'.